Refactor things a bit
