+++ src/quartic/__init__.py
@@ -2,3 +2,6 @@ from .ipython_magics import *
 
 from .quartic import Quartic
 from .utils import QuarticException
+
+from .runner.step import step
+from .runner.dataset import writer
